Fix off_t and size_t printf warnings on 32/64 bit systems.
Dieharder already uses _FILE_OFFSET_BITS=64, so off_t type
is always 64bit signed integer.

However, on 32bit platforms is is presented as long long,
while on 64bit it is just long (as long type is 64bit there).

As there is no specific format prefix for off_t, IMO, the safest
is to use integer promotion - use larges type prefix %lld
and cast value to (long long) in printf.

Also switch printf to signed integer, as off_t is always signed.

For size_t we have the same issue, but here we have already
print prefix %zu that will handle size_ on all patforms.

+++ dieharder/output.c
@@ -83,12 +83,12 @@ void output(Dtest *dtest,Test **test)
  if(strncmp("file_input",gsl_rng_name(rng),10) == 0){
    if(tflag & TFILE_RNG_STAT) {
      if(strncmp("file_input_raw",gsl_rng_name(rng),14) == 0){
-       fprintf(stdout,"# The %s %s %lu rands (%llu bytes) were used\n",
-               filename, gsl_rng_name(rng),file_input_get_rtot(rng),
+       fprintf(stdout,"# The %s %s %lld rands (%llu bytes) were used\n",
+               filename, gsl_rng_name(rng),(long long)file_input_get_rtot(rng),
                (unsigned long long)file_input_get_rtot(rng) * sizeof(uint));
      } else {
-       fprintf(stdout,"# The file %s %s %lu rands were used\n",
-               filename, gsl_rng_name(rng),file_input_get_rtot(rng));
+       fprintf(stdout,"# The file %s %s %lld rands were used\n",
+               filename, gsl_rng_name(rng),(long long)file_input_get_rtot(rng));
      }
      if(file_input_get_rewind_cnt(rng) == 0){
        fprintf(stdout,"# The file %s was rewound %u times\n",
